Added policy for Documents to check roles

+++ app/Http/Controllers/DocumentController.php
@@ -44,6 +44,9 @@ class DocumentController extends Controller
      */
     public function create()
     {
+        if (auth()->user()->cannot('create', $this->document)) {
+            return redirect('home')->with('alert.danger', 'You do not have access to upload documents.');
+        }
         return view('documents.create');
     }
 
@@ -55,6 +58,10 @@ class DocumentController extends Controller
      */
     public function store(Request $request)
     {
+        if (auth()->user()->cannot('create', $this->document)) {
+            return redirect('home')->with('alert.danger', 'You do not have access to upload documents.');
+        }
+
         // Store the document in Storage
         $file = $request->file_path
             ->storeAs('documents', $request->name .'.'. $request->file_path
@@ -97,6 +104,10 @@ class DocumentController extends Controller
      */
     public function destroy($id)
     {
+        if (auth()->user()->cannot('delete', $this->document)) {
+            return redirect('home')->with('alert.danger', 'You do not have access to remove documents.');
+        }
+
         $document = $this->document->where('id', '=', $id)->first();
         unlink(public_path("storage/". $document->file_path));
         $document->delete();
